// apps/web/src/http/uazapi/types.ts
export interface FindMessageItem {
  id: string;
  messageid: string;
  chatid: string;
  sender?: string;
  senderName?: string;
  isGroup?: boolean;
  fromMe: boolean;
  messageType: string;
  source?: string;
  messageTimestamp: number;
  status?: string;
  text?: string;
  quoted?: string;
  edited?: string;
  content?: any;
  fileURL?: string;
  track_source?: string;
  track_id?: string;
}